refactor: make drawing-block fully self-contained and decompose controller
- Move drawing types, actions, apply, and history into drawing-block/
  with its own Brand utility and branded IDs (zero external imports)
- Extract ActionRecorder (~85 lines) from controller for action log
  and undo/redo cursor management
- Extract generic typed Emitter (~17 lines) from controller
- Slim down DrawingBlockController from 247 to ~130 lines as orchestrator
- Delete backward-compat shim layers (src/drawing/, src/types/drawing.ts)
- Update all imports to reference drawing-block/ directly
- Add ARCHITECTURE.md, CLAUDE.md, .cursorrules, expand README.md
- Add doc-update reminder comments to all source files
