+++ CLAUDE.md
@@ -1,0 +1,36 @@
+# Project Rules
+
+## Documentation Sync
+
+When any code changes affect the module structure, public API, dependency graph, or design patterns:
+
+1. Update `README.md` (project structure, features, examples)
+2. Update `ARCHITECTURE.md` (module graph, design decisions, testing strategy)
+3. Review inline doc-update comments in changed files
+
+This applies to: adding/removing/renaming modules, changing exports, modifying class responsibilities, altering the dependency graph, or changing architectural patterns.
+
+## drawing-block is self-contained
+
+The `src/drawing-block/` module must have **zero imports from other `src/` modules**. It owns its own types, branded IDs, and all drawing logic. This is intentional so it can be extracted as a standalone package. If you need a type from drawing-block elsewhere, re-export it (as `types/ids.ts` and `types/drawing.ts` already do).
+
+## Code Style
+
+- SOLID principles: single responsibility classes, depend on interfaces not implementations
+- DRY: extract shared patterns into utilities (e.g., `Emitter`, `ActionRecorder`)
+- Immutable data: use `readonly`, return new objects from state transitions
+- Branded types for all IDs (use `Brand<string, "Name">`)
+- Pure functions for data transformations (apply, history)
+- Discriminated unions for polymorphic types
+
+## Commit Messages
+
+- NEVER include "Claude", "Copilot", "Co-Authored-By", "Generated by", or any AI/coding agent attribution in commit messages, trailers, or anywhere in the git metadata.
+- Keep commit messages clean and human-authored in style.
+
+## Testing
+
+- Unit tests live alongside source files (`*.test.ts`)
+- E2E tests live in `e2e/`
+- Run `npm test` for unit tests, `npm run test:e2e` for e2e
+- All tests must pass before merging
