+++ src/drawing-block/controller.test.ts
@@ -1,11 +1,10 @@
 import { describe, it, expect, vi } from "vitest";
 import { DrawingBlockController } from "./controller.js";
 import { DefaultDrawingEngine } from "./default-engine.js";
-import type { DrawingScene } from "../types/drawing.js";
-import type { DrawingAction } from "../drawing/actions.js";
-import type { DrawingHistoryState } from "../drawing/history.js";
-import { createDrawingId, createElementId, createHistoryGroupId } from "../types/ids.js";
-import type { DrawingElement } from "../types/drawing.js";
+import type { DrawingScene, DrawingElement } from "./types.js";
+import type { DrawingAction } from "./actions.js";
+import type { DrawingHistoryState } from "./history.js";
+import { createDrawingId, createElementId, createHistoryGroupId } from "./types.js";
 
 let groupCounter = 0;
 const makeGroupId = () => createHistoryGroupId(`hg-${++groupCounter}`);
