--- src/drawing/index.ts
@@ -1,3 +1,0 @@
-export * from "./actions.js";
-export * from "./apply.js";
-export * from "./history.js";
