+++ src/index.ts
@@ -1,6 +1,5 @@
 export * from "./types/index.js";
 export * from "./body/index.js";
-export * from "./drawing/index.js";
 export * from "./drawing-block/index.js";
 export * from "./comments/index.js";
 export * from "./adapter/index.js";
