+++ src/types/index.ts
@@ -1,4 +1,3 @@
 export * from "./ids.js";
 export * from "./blocks.js";
 export * from "./document.js";
-export * from "./drawing.js";
